lorgnette: add AllScanCombinationsTest()
This test iterates through each combination of color mode and resolution
advertised by a scan source. The scanned images are sanity-checked to
ensure they are the correct size for the given resolution, and also the
correct color space.

When run on an ADF source, the test will prompt the user to put paper in
the ADF and enter the number of pages used before each scan is
performed.

BUG=b:191933619
TEST=FEATURES=test emerge-nami lorgnette-wwcb-tests

Change-Id: I976e1303d360e06fbd9c75a5e83694f11c4fb14c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3429690
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>