debugd: Adds the perf_logging feature and flag
Adds the perf_logging flag and implements the functionality for
measuring and logging the time it takes for GetBigFeedbackLogs sub-tasks
to complete. This is done by adding a Lap function to the Stopwatch
object, which records the elapsed time since the stopwatch was
instantiated. Makes reporting to UMA optional, so that a Stopwatch can
be created for measuring and logging performance locally only.

BUG=b:191179515
TEST=Log messages are written to /var/log/messages* files correctly and
are not written when the flag is not supplied.
TEST=All debugd tests pass.

Change-Id: I73e259b60cb8c1bbb5f8ab03c82e22ee60863675
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3430582
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>