shill: converge on error reporting via callbacks in Start/Stop
This removes the dual-error-reporting pathway in Start and Stop
calls on devices, which was leading to circumstances where the unit
tests would only test code that was unused in production, and avoid
testing the parallel codepaths that production code actually used. As
such, we were actually missing a lot of test coverage.

BUG=b:172215298
TEST=unit tests, wifi_matfunc, cellular_cq

Change-Id: I35b43cf7746ce518854c79a6b1f6485a9fcb8467
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3433307
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>