shill: Add WiFiSecurity class
So far shill keeps only one member for "security" (std:string security_)
which is either Security or SecurityClass depending on the value.  This
is a bit inconvenient, so the intention is to split these two properties
and keep them separately.  The WiFiSecurity class that is introduced is
meant to keep the Security property and the purpose of keeping it as
a separate type is twofold:

- make it easier to figure out what property should be used with the
  help of compiler (I want a typeful difference between Security and
  SecurityClass properties so that all places in code that use Security
  or SecurityClass are updated properly)
- new Security property should implement logic allowing to combine
  multiple endpoints (see go/cros-wifi-finegrained-security) and this
  is implemented inside WiFiSecurity class

BUG=b:226138492, b:207588328
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: I58b140109ce8c294e83c7332e11502cf593dd9ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3470055
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>