secanomalyd: Report processes in the init PID namespace.
In order to aid analysis of anomalous systems, report the list of
processes in the init PID namespace. This provides enough information
while avoiding processes that may contain PII, like Android app
processes (which run in a non-init PID namespace or inside a VM.)

This requires granting secanomalyd the CAP_SYS_PTRACE capability to
allow secanomalyd the ability to inspect the PID namespace of
processes. This is not super exciting, but if we eventually determine
that we don't want to keep this capability we can always find a
different way to filter processes.

BUG=b:216855846
TEST=Deploy, run, check crash report at /var/spool/crash.
TEST=See list of processes.

Change-Id: I328ba78efe3e43368d58b3c2f8e95168d96f497c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3437845
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>