missive: Implement disconnectable client
This is a wrapper on top of an abstract client which may become
available or unavailable at any moment. Calls made through the client
are rejected with error if the client is not available, and cancelled
immediately returning an error, if it becomes unavailable while the call
is in progress.
The wrapper is intended to be used for dBus connection to missive client
in Chrome and from daemon to Chrome in missived.

BUG=b:205814822
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I9d0bea844454822ef48fd38f2d2ab4c11bf31019
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3438146
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Reviewed-by: Hong Xu <xuhong@google.com>
Reviewed-by: Zach Trudo <zatrudo@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>