chromeos-config: Derive the legacy PSU type solely from the DEVICETYPE
For the limited number of remaining pre-unibuilds, the PSU type
happens to entirely be a function of DEVICETYPE: that is, for a
DEVICETYPE of CHROMEBOOK, the PSU type is "battery".
Otherwise (DEVICETYPE is CHROMEBOX), the PSU type is "AC_only".

This avoids the need to shell out to mosys.

BUG=b:217920839
TEST=unit tests pass

Change-Id: Ib841441b8bb19616c2db9282e3538ae9c8d7452c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3439570
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>