cryptohome: Fix AddCredential for initial keysets
Fix the error in the AddCredential API that it would create multiple
initial keysets if called multiple times. Internally, the fix changes
that this function doesn't look at the "user_exists" flag anymore,
since it wasn't the right signal for checking.

This commit, besides fixing this issue, also is preparation for
changing the "user_exists" flag to be dynamically updated when the user
is created. This will be important for enforcing the correct semantics
of the CreatePersistentUser and AddAuthFactor APIs in the future.

BUG=b:217971798
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Ib0e2f9538ea977120480f74729a57ad8426dd272
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3441735
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>