minios: Destruct previous_screen when we return from an error_screen
The previous_screen will respond to events that are not proxied
through the screen_controller. This will result in two screen objects
drawing to the screen at the same time. Fix this by destructing the
previous_screen when we transition from an error screen to a newly
allocated screen.

Also set `ScreenNetwork` to dropdown closed state when we transition
to an error to disable drawing to the screen due to NetworkManager
events.

BUG=b:199369691
TEST=FEATURES=test emerge-$B minios

Change-Id: Ie2d5514f9700360520b8d259afc14cd76de1fe70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3444762
Tested-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>