chromeos-config: add schema option for NNAPI drivers
Currently, the mechanism for selecting which drivers NNAPI should use is
determined at runtime through the use of an environment variable (see
src/aosp/frameworks/ml/chromeos/driver_loader.cpp) that is currently
only set (as far as I can tell) by manually setting it to some value on
your test device. This mechanism is fine for testing, as currently it
defaults to a driver that works on all devices, but it does not allow us
to automatically be sensitive to particular board architectures and
enable the most ‘appropriate’ driver.

Utilise the flexibility of chromeos-config to configure, on the model
level, what drivers are most appropriate to use.

BUG=b:196007189
TEST=Built and deployed on asurada. Unit tests run (FEATURES=test sudo
-E emerge chromeos-config-host)

Change-Id: I6abda646ff44dbecbe28f20e9378771dc564ee3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3445560
Reviewed-by: Michael Olney <mpolney@google.com>
Reviewed-by: Jack Shen <jackshen@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Tom Shafron <shafron@google.com>
Tested-by: Tom Shafron <shafron@google.com>
Commit-Queue: Tom Shafron <shafron@google.com>