shill: vpn: Fix wrong usage of (New)L2TPIPsecDriver
The current implementation of the logic "NewL2TPIPsecDriver should be
used if tunnel group is not set" is not correct. Currently, if the
service is load from the profile (i.e., not configured directly from
Chrome), the driver will actually be picked up randomly, based on
whether the bool value |use_new_l2tp_driver| is zero-initialized.

BUG=b:204261554
TEST=manually checked that the script can control which driver is used
  properly.

Change-Id: I7f2201d8163184a541f2f5a8b399abf692ed178c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3447730
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>