hammerd: init: Fix unexpected hammerd.conf abort
Inside the script stanza of a .conf script, the commands will be
executed using `sh` and the "-e" option is always used, so any command
that fails will terminate the script.

In addition, cros_config exit with code 1 when it fails to query the
key. This means that any board that does not support querying detachable
base information via cros_config will fail to run hammerd.conf in the
end, for example Soraka.

This fixes the issue by returning true when the querying commands fail,
and also slightly improves the code quality and adds more log prints.

BUG=b:218447539
TEST=emerge-$BOARD hammerd
     cat /var/log/hammerd.log and verify hammerd.conf works on Soraka

Change-Id: I5d47a233ae0658b8aaab01af000390b56c3a2ccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3450157
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>