chromeos-config: Remove GetNamedTuple helper
This appears to be syntactic sugar used by just a few tests, so remove
it for consistency.

Note that if we want to reintroduce this syntax again (and I'm not
saying we should), we should try and uniformly adopt it (instead of in
just a handful of tests) and use a SimpleNamespace instead of a
NamedTuple, as the former does not require dynamic singleton type
creation.

BUG=none
TEST=unit tests pass

Change-Id: Ia75f3295d01cfb67bcd198df5618bbf8d25ac568
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3456885
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>