cryptohome: Don't do ChallengeCredential mount with active session
DoMount() function doesn't check for an active session if the key type
is challenge credential. This doesn't introduce any problem in the
current code as we don't expect any mount call with challenge credential
key type if there is an active session. Also even after we add the
authenticate AuthSession with challenge credential key type and start
the mount operation with that challenge credential  authenticated
session there won't be any problem, because expected API changes won't
let that. But the existing code would introduce a confusion for the
reader when we add challenge credentials to AuthSession. So make the
challenge credential and AuthSession decisions mutual exclusive as
in other key types.

BUG=b:215200429
TEST=Builds and unit tests pass.

Change-Id: I4f39e2cc1f0f4480ff78989c935b895c7d0097ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3459498
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>