crosh: Add arg to --dev flag
Support an optional argument for the --dev flag. --dev=true is the same
as --dev, and --dev=false forces crosh to not load dev commands. This is
useful when running tests to ensure that a command is or is not present
in normal mode.

BUG=b:215171945
TEST=Check "shell" command with --dev flag set to true and false

Change-Id: Idd9ba5aba97c32e9d1ef1a88907f7b3e15c54e4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3516534
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>