shill: client: Don't call RemoveObjectProxy on dtor
RemoveObjectProxy synchronously removes the tracking set of
the D-Bus rules used, but then asynchronously actually cleans
them (inside D-Bus thread). This creates a race condition
on cleanup when the client is removed right before the D-Bus
connection is shutdown. On destructor, the owner of the D-Bus
connection should be responsible for the cleanup (similar to
the pattern of other clients).

The scenario is as follows:
1. RemoveObjectProxy removes the tracking set.
2. D-Bus shutdown sees nothing in the tracking set and does
   not clean the rules.
3. D-Bus connection is shut down and the asynchronous clean
   up is not run.

BUG=b:219648324
TEST=Shutdown ARC; DCHECK on bus.cc is not hit
TEST=units

Change-Id: I323babe6a5efe5c1a3acef899cf9605c63d9fc71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3468795
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>