permission_broker: allow Spikerbox USB device
https://backyardbrains.com/products/spikerbox

This device is currently presenting as an HID to keep end-users
from needing to install drivers (as they may lack the necessary
admin privileges). Given that the device actually only has vendor
usages it should be safe to allow to be attached to the guest.

Ideally we could check for HID devices that only have vendor
usages and allow those more generally, but this would be a
quick fix for the user until more time can be devoted to the
issue.

BUG=chromium:1290884
TEST=tested with spikerbox neuron pro on volteer device

Change-Id: I5aee52682f48443134b345057fa139fc0e4efaf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3472618
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Maciek Swiech <drmasquatch@google.com>
Commit-Queue: Maciek Swiech <drmasquatch@google.com>