patchpanel: ndproxy: remember src MAC address of guests
This patch introduces in NDProxy a map of neighbor MAC addresses for all
neighbors seen on guest interfaces monitored by NDProxy. This allows to
consistently obtain the correct destination MAC address when forwarding
a packet from the network to the guests, as it is observed that
sometimes querying the kernel neighbor table does not work.

Tested ARC IPv6 connectivity on both trogdor (QC modem) and hatch
(Fibocom L850) when cellular is the default network:
  - test-ipv6.com reports IPv6 in Firefox App running in ARC
  - adb shell dumpsys wifi tools reach ipv6.google.com succeeds
  - adb shell dumpsys wifi tools http https://ipv6.google.com succeeds

Tested Crostini IPv6 connectivity on the same DUT setups as above:
  - curl ipv6.google.com works in both the user lxd container and the
    Termina VM shell.
  - ip -6 addr ls shows IPv6 provisioning in both the lxd container and
    the Termina VM shell.

When logging out of the current session and restarting a new session,
all of the above keep working.

BUG=b:187918638
TEST=FEATURES=test emerge-$BOARD patchpanel. Plus manual tests above.

Change-Id: I69619c199bc144c9d3da7438219881da3f34ca05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3480675
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>