shill: vpn: Add PassphraseRequired property in IKEv2Driver
This patch adds PassphraseRequired property in the Provider dict for an
IKEv2 VPN service. This property returns whether PSK or the password for
EAP is not set. Since the credential fields (e.g., password) are not
readable via RPC, this property will be used (by Chrome) to check if the
service is connectable based on whether the credential fields are filled
or not.

BUG=b:210064468
TEST=unit_tests

Change-Id: I6a63e9bc843ea0788dc343c93834dba937826827
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3476757
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>