u2fd: Set max iterations for u2fhid_fuzzer
Currently we will use up all bytes in a single fuzzing round. However,
data size might be as large as millions of bytes, resulting in tens of
thousands of iterations, exceeding timeout for a single round in
ClusterFuzz. Set a max iteration (100) for it.

BUG=b:217652224
TEST=run testcase in the bug

Change-Id: I169df25ca8ec7f3f226d69c672b3018b64fd9c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3476761
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>