missive: Handle reused response_callback_.
It looks like the reported crash is caused by an attempt to call
response_callback_ after it was already called (There is a DCHECK there,
but it is inactive in official builds, and doing if instead is more reliable
while doing no harm).
It probably happens because callback provided to delegate was not always
called - this is also fixed here.

BUG=b:220894570
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I4e1e943a97003590ca2cb4ac2082eeda937b702b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3482180
Reviewed-by: Zach Trudo <zatrudo@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>