cryptohome: Add Encrypt/Decrypt with KeyBlobs
Add the EncrypEx() and DecryptEx() with the KeyBlobs. These functions,
on contrary to the existing Encrypt() and Decrypt() functions take the
KeyBlobs and AuthBlockState as a parameter and leave the derivation to
the caller. This is needed to take the AuthBlocks out of VaultKeyset.
After all the code connections are completed existing Encrypt/Decrypt
functions will be removed.

BUG=b:200147792
TEST=Unit tests pass.

Change-Id: Ie0e781934800f21086b87f1a7f95b8e47cf80bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3484804
Reviewed-by: Zi Lin <lziest@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>