debugd: Reduce arbitrary sleep time in log tool
Reduces an arbitrary sleep time for CreateConnectivityReport function
from 5 seconds to 3 seconds. Based on measurements, shill takes about 10
milliseconds to run a connectivity test on an interface. Additionally,
there is already many seconds of buffer between when the call to shill
is made and when the reports are sent, which means the call has much
more time to complete in practice. In the worse case shill takes ~10
seconds to complete a connectivity test (no upstream connectivity).

BUG=b:191179515, b:220896175, chromium:1218182
TEST=All debugd tests pass. System information loads 2 seconds faster
when creating feedback report and the network test information is
complete.

Change-Id: Ied77278665e418817691b05a8a68c104bd864798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3470091
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Auto-Submit: Shervin Oloumi <enlightened@chromium.org>