permission_broker: allow detaching Arduino Uno Wifi Rev2
This is a serial (USB CDC ACM) device that was not in the existing list
of Arduino device IDs. Allow detaching it from the kernel serial drivers
so it can be used with Crostini.

Vendor and product ID for this device can be found here:
https://raw.githubusercontent.com/arduino/ArduinoCore-megaavr/HEAD/boards.txt

BUG=chromium:1248796
TEST=share Arduino Uno WiFi Rev2 with Linux

Change-Id: Ia67ced807fdb4a2ee4cff5071f19fcdd3c90dc42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3484886
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>