power: Use custom default proximity state
On some boards the proximity state is near by default,
which causes the modem to be stuck in low power state.
This change allows the user to configure the default
proximity state on a per sku basis in cros-config.
Use set-default-proximity-state-high config to set
the default proximity state.

BUG=b:213141895
TEST=emerge-dedede power_manager
     cros deploy power_manager
     restart powerd or reboot the device
     Verify the output power is 22db

Change-Id: I5508effdebff59da1d65795445188e1551eb1577
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3484887
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Madhav <madhavadas@google.com>
Tested-by: Madhav <madhavadas@google.com>