ml: Retroactively fix versioning in model.mojom
**This CL will need to be chumped** in order to bypass the stable
mojom presubmit.**

Before a presubmit had been guarding it, some
non-backwards-compatible changes were made to [Stable] interfaces
in this mojom. Then an appropriate versioning information was added
allowing the changes to be synced to the Chrome side where
the presubmit exists. However, an erroneous version was assigned,
so that GpuDelegateApi needs to be nailed down.

BUG=chromium:1283517
TEST=1) build_packages && cros_workon_make chromeos-base/ml --test
TEST=2) Run a dry run

Change-Id: I78d31b2ef5e54cf2561c638432617b0ca8109678
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3487566
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Dmitry Grebenyuk <dgrebenyuk@google.com>