shill: remove unused shill-cli code and package
The original author left Google before this effort could yield a
functional dev tool. This package and code should be removed since
nobody is maintaining it and all Connectivity engineers use direct dbus
commands or platform2/shill/test-scripts on a day-to-day basis.

BUG=b:172222930
TEST=The output of $ equery-$BOARD depends shill-cli shows that nothing
depends on this package. It is currently not being build.

Cq-Depend: chromium:3488246
Change-Id: Ie7c9104dcee2848d87f2d2b6aec6f5b7e256f1cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3488306
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>