vm_tools: concierge: add ListVms method
There is no way to query which VMs are running in a general way. Each VM
might have a vm-specific way. For example ARCVM has IsArcVmEnabled, but
that does not take suspend/lazy start into account. Borealis has a
different method. Termina VMs can have custom names, etc.

It is possible, with VM specific codepaths, to enumerate VMs in Chrome,
but it's complicated, and not available outside of Chrome.

It would be nice to have a method to enumerate VMs for the following
reasons:
* Tast tests can get accurate VM names
* Crash reporters and UMA metrics can easily add metadata about which
  VMs are running

This CL introduces a ListVms method to concierge that returns the VMs in
concierge's VmMap that match a passed cryptohome_id. All the same data is
already sent with the existing VmStartedSignals, but now it is also
possible to query.

BUG=b:214484991
TEST=on dut, start ARCVM and crostini, concierge_client --list_vms shows
both running.

Change-Id: I8acc5010804804539cd990cfceebb3b63a3c4a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3493812
Reviewed-by: Alexandre Marciano Gimenez <raging@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Charles William Dick <cwd@google.com>
Commit-Queue: Charles William Dick <cwd@google.com>