power: Don't disable ramoops console
b/63737106 introduced a workaround for Intel platforms where the serial
console was preventing the system from entering S0iX. The workaround
didn't take into account the `ttynull` console. This console gets added
by default when no other console is available. This has the unintended
consequence of disabling the ramoops console and thus making it
difficult to debug suspend failures.

    # cat /proc/consoles
    ttynull0             --- (EC    )  247:0
    ramoops-1            -W- (E  p a)

Since this is an Intel specific work around it would be nice to limit it
to Intel only. That can be done as a follow up. I'm not sure how best to
do this though.

BUG=b:219973337, b:175428322, b:63737106
TEST=Suspend guybrush and cause kernel panic on suspend. Verify ramoops
was generated.

Change-Id: I35ba326716abbfa8737d8fb0b9325a3482169e31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3495705
Reviewed-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Tested-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>