power: add mt7921s driver to set_wifi_transmit_power.cc
set_wifi_transmit_power.cc contains a map of drivers to vendors which
allows the daemon to use different implementations for different
vendors. pico6 uses the mt7921s driver, which is not currently
represented in this map. Add this driver so the daemon knows to use
the MTK implementation for this driver.

BUG=b:221930519
TEST=tast run ${DUT} tast.wifi.SetTXPower
Passes on pico6

Change-Id: Iff671a39c264361eab167188ddf984fa811e0965
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3498398
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>