concierge: Add Vsock Connection to ARCVM's LMKD
To better respond to quickly changing memory pressure situations in
ARCVM, concierge will maintain a vsock connection to LMKD. When LMKD
thinks it needs to kill, it will ask concierge to confirm this kill.
Based on the priority of the candidate process, concierge will either
deflate the virtio balloon of ARCVM to make room for the process, or
instruct LMKD to proceed with the kill.

BUG=b:210075795
TEST=CQ, cros_run_unit_tests --board volteer --packages
chromeos-base/vm_host_tools, manual local integration testing with
LMKD changes

Change-Id: I375891b0f7ac6155ce67c23bf1e434647d1c0ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3510915
Commit-Queue: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Tested-by: Kameron Lutes <kalutes@chromium.org>