camera: use cached TET and HDR ratio for the initial frames
When we first open a camera device, we queue several frames in a batch
to fill up the pipeline. These frames are queued before we get the first
AE stats, hence Gcam AE does not set the exposure values in their
capture requests and we get black frames. Let's use the cached TET and
HDR raio, if available, in the bootstrapping frames, so that the AE
transition is more natural.

BUG=b:221166548
TEST=Manually with CCA and verify that the brightness of the camera
preview does not flicker when switching modes.

Change-Id: I3f52bd9269ebb0b6025523f5ff625a50f9af0489
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3499802
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>