hps: update stage0 if version is different
Drop explicit invocation of gpioset, because the kernel manages the HPS
peripheral power state now. The hps-factory command will unbind and
re-bind the kernel driver, resetting the peripheral.

Use the new --check-version option for `hps-factory write-stage0`, to
ensure that devices in the field will have their stage0 upgraded when we
bump the stage0 version. This gives us a way to roll out the MP key to
existing non-production devices which are using the dev key.

BUG=b:222000200
TEST=start hpsd  # blank stage0, it writes it
TEST=start hpsd  # mismatched stage0 version, it erases and rewrites it
TEST=start hpsd  # matching stage0 version, it leaves it alone

Cq-Depend: chromium:3505286
Change-Id: I417e4f4bed5971ef756b42a3db71b220abc102eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3501433
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>