cryptohome: Prepare WebAuthn secret only when necessary
Use the new parameter `unlock_webauthn_secret` in CheckKey to determine
whether it needs to prepare the WebAuthn secret for u2fd to fetch. Also,
the secret is actually not needed to be stored when mounting, only its
hash is needed. Add a new method PrepareWebAuthnSecretHash to be used
when mounting. These changes guarantee that the WebAuthn secret is only
derived when it will be used. Modify the tests and cryptohome cli
accordingly.

BUG=b:222439151
TEST=cryptohome unittests; tast run $DUT u2fd.*

Change-Id: I4b2d3040cd924060b57b64c002fb723d673c3a43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3499908
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>