init: Move dev_image update to post logical volume mount
If the dev-image is mounted from a logical volume, we will
fail to update the dev-image utils. Instead, do the update post
mounting the logical volume.

Additionally, account for the fact that dev_image is a mount point,
instead of moving the mount point around, move it's contents.

BUG=b:221493224
TEST=CQ

Change-Id: I96d6cf24f501d7e7e903f775558a0320202075bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3501121
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>