camera: Try subscribe privacy state monitor when stream on
Originally it is subscribed when the device is connected. However,
something change and now when trying to query the privacy button state
at that timing it will fail with "Device resource busy" message.

This CL moves the timing later so that it will try to monitor it once
the camera is streamed on.

BUG=chromium:1240725
TEST=Manually tested on Jinlon

Change-Id: I2d1483d752814e400135f839b0088c46089a7854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3505480
Tested-by: Wei Lee <wtlee@chromium.org>
Auto-Submit: Wei Lee <wtlee@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>