diagnostics: Modify NVMe self test routine progress update logic
Currently two conditions must be met for the routine to be determined as
running:
- the percentage < 100%
- the routine status is running

However, on certain devices, the routine status will still be running
even if the percentage hits 100, which means the percentage should not
be used to determine its status. Modify the progress update logic so
that only the routine status is used.

BUG=b:175305207
TEST=FEATURES=test USE="-cros-debug" emerge-${BOARD} diagnostics
TEST=tast run $dut health.DiagnosticsRun.nvme_self_test

Change-Id: Ic75a409edc8d65eefce654424799cb95dfb9a909
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3506264
Tested-by: Denny Huang <dennyh@google.com>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Commit-Queue: Denny Huang <dennyh@google.com>