shill: cellular: Check for MVNOs without filters
If a MVNO does not have a filter, it is always selected. This is very
dangerous when making modb changes, as forgetting the filter will make
the MVNO the only MVNO for a MNO, and it will also override the MNO
data.

BUG=None
TEST=Run unit tests and fix existing issues.

Change-Id: I7f29d6da6f7f343d0ca79ca2b46911df80a566f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3507397
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>