shill: PortalDetector uses ManagerProperties directly
This patch removes PortalDetector::Properties and instead makes
PortalDetector directly use the ManagerProperties. The Properties class
defined by PortalDetector was not adding any particular logic but was a
source of boilerplate code for passing URL properties to the portal
detection logic.

Using directly the ManagerProperties greatly facilitates introducing a
new set of URL properties in the follow-up patch.

This patch also slightly changes the portal detector URL fallback logic
such that the default URL can still be selected in retry attempts. This
increases the retry pool of URLs and increases the chance of portal
detection when a retry is caused by a genuine network error.

BUG=b:211563221
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I7feaed75d876c3f3e3f486743e9abb8124a4be0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3513678
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>