shill: cellular: Set default values for authentication and ip_type
Currently, shill does not have a default value for authentication or
ip_type when a value is not present in the database. This makes the
assumption that ModemManager knows how to handle this correctly, but
this is not the case for every modem, as each modem has a different
default value.

BUG=b:222563263
TEST=Run APN tests on callbox

Change-Id: Ia586ae4dd179e7b5214b523c61394af8a63ce52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3516552
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>