cryptohome: Add a flag to negate fscrypt v2
We want to have a version update vault compat tests. That includes v1/v2
compat. Presently it is impossible to create a v1 vault on v2 supporting
boards, even though v1 vault still should work. the flag is to be passed
in tests with CRYPTOHOMED_ARGS to negate enabled v2 in order to create a
v1 vault.

BUG=b:224366279
TEST=builds

Change-Id: I9ef7cdab23d8e53cedb4bb80ae93a8787f79aec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3517186
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>