camera: Fix crash when updating option for auto framing (part 2)
After CL:3343576, the auto framing stream manipulator configure
streams differently when the feature is enabled or disabled, causing the
cros_camera_service to crash when the enable setting is toggled at
runtime due to mismatch of response buffer with the streams.

To solve this, we instead always pass the full_frame_stream_ to HAL as
an additional stream instead.

If auto framing is disabled, we simply remove the output buffer of
full_frame_stream_ from the response. And if auto framing is enabled, we
ignore the original output buffer from HAL for other streams and crop
them from full_frame_stream_ as before.

BUG=b:223739596
TEST=change /run/camera/auto_framing_config.json and check auto framing
     is correctly toggled.

Change-Id: Ibfc14c4c44004b2871522338bdab5da0da140b51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3518709
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>