missive: Ensure TestEvents run on the main thread
Flakes that were detected recently seem to be related to Write operation
completion guarantee. RunLoop::Quit is allowed to run on threads, but
in that case its completion is not immediate, and  waiting for the
result on the main thread could be delayed.
Fixing it here by updating result on the main thread (the same one where
the TestEvent was placed).

BUG=b:222539905
BUG=chromium:1306208
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I799d6f53f5638ab1d67b21af1c1ce96dfb2b7ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3523727
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>