system_api: Add an EnsureLaunched request to the vm launch service
This request is used to launch VMs via chrome, as per go/vmc-launch
phase 2.

Rather than requiring users to construct the increasingly complex `vmc
start` invocation with all of its args, `vmc launch` is used to ask
chrome, which is the authority on the correct set of args, to launch a
vm for it.

BUG=b:224660520
TEST=dbus-send --print-reply --system --type=method_call \
     --dest=org.chromium.VmLaunchService /org/chromium/VmLaunchService \
     org.chromium.VmLaunchService.EnsureVmLaunched

Change-Id: I332931c1e8a0c5baad103de2650d5e0bd2fb3f1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3524521
Auto-Submit: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>