patchpanel: ndproxy: do not autodetect irregular RA networks
This patch removes the autodetection of upstream interfaces with
"irregular RAs" as described in b/187918638. Since crrev/c/3499601,
ndproxy is now always told directly on which interfaces are irregular
RAs expected (i.e Cellular interfaces).

This prevents a bug where a regular physical network interfaces like
WiFi or Ethernet could be classified as interfaces with irregular RAs
when the gateway sending an RA is not known yet as a neighbor. This bug
can be triggered consistently by:
  - disabling autoconnect for a dual stack or IPv6 only WiFi network
  - rebooting the DUT and starting a user session
  - manually connecting to the WiFi network

BUG=223863095
TEST=FEATURES=test emerge-$BOARD patchpanel. Also verified:
  - ARC and Crostini IPv6 connectivity works consistently on
  GoogleGuest-IPv4 WiFi
  - Cellular interface is added to list of irregular RA networks.

Change-Id: I1d7c51bc979791df6b71f56a7b34a31df12f761d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3524522
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>