vm_tools: Modify `vmc launch` to talk to chrome
See go/vmc-launch section "Phase 2". The intention is to allow chrome to
be the authority on the "correct" way to launch a VM. `vmc start` will
still exist and be used by developers, but users who prefer command-line
should be encouraged to switch to `launch`.

BUG=b:200896773
TEST=vmc launch {borealis,pluginvm,termina,crostini}

Change-Id: Ib052b95b878180d9bbdff607a850958219277c47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3529048
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>